appease numpy

+++ test/test_converter.py
@@ -19,7 +19,7 @@ def array_from_dict(data):
 def test_array_converter():
     @xattree
     class TestClass:
-        a: NDArray[np.integer] = array(converter=array_from_dict)
+        a: NDArray[np.int_] = array(converter=array_from_dict)
 
     obj = TestClass(a={0: 1, 2: 3})
     expected = np.array([1, 0, 3])
@@ -33,7 +33,7 @@ def test_array_converter_takes_self():
 
     @xattree
     class TestClass:
-        a: NDArray[np.integer] = array(converter=attrs.Converter(convert, takes_self=True))
+        a: NDArray[np.int_] = array(converter=attrs.Converter(convert, takes_self=True))
 
     obj = TestClass(a=1)
     expected = np.array(2)
@@ -47,7 +47,7 @@ def test_array_converter_takes_field():
 
     @xattree
     class TestClass:
-        a: NDArray[np.integer] = array(converter=attrs.Converter(convert, takes_field=True))
+        a: NDArray[np.int_] = array(converter=attrs.Converter(convert, takes_field=True))
 
     obj = TestClass(a=1)
     expected = np.array(2)
@@ -62,7 +62,7 @@ def test_array_converter_takes_self_and_field():
 
     @xattree
     class TestClass:
-        a: NDArray[np.integer] = array(
+        a: NDArray[np.int_] = array(
             converter=attrs.Converter(convert, takes_self=True, takes_field=True)
         )
